Refactor for better OOP
